requirements.txt
paho-mqtt==2.1.0
grpcio==1.76.0
asyncio==4.0.0
nicegui==3.1.0